+++ src/main/java/cz/sizi/bikeo/service/impl/RoleServiceImpl.java
@@ -39,7 +39,7 @@ public class RoleServiceImpl implements RoleService {
 	}
 
 	@Override
-	public Role findById(Integer id) {
+	public Role findById(Long id) {
 		return roleDao.findById(id);
 	}
